fixed breakages on sfml2 as on 14th
git-svn-id: http://projects.hashstash.in/svn/trivial@29 659a0b78-bc15-4dc9-aeea-a5f01ff3c5d0

+++ src/TrivialCamera.h
@@ -3,6 +3,7 @@
 
 #include <iostream>
 #include <string>
+#include <math.h>
 
 #include "SFML/Graphics.hpp"
 
@@ -41,7 +42,7 @@ public:
     float getCenterX() const { return _x + _halfWidth; }
     float getCenterY() const { return _y + _halfHeight; }
 
-    void setZoom(const float &percent = 100) {
+    void setZoom(const int &percent = 100) {
         /*
          simple, 100 = 1; 150 = 1.5; 50 = 0.5
          but if we were to set the zoom to 1.5 it would show 150% of the screen
